Removed deprecated utility classes.
