+++ lib/Auth/AWS.php
@@ -2,7 +2,8 @@
 
 namespace Sabre\HTTP\Auth;
 
-use Sabre\HTTP\Util;
+use Sabre\HTTP;
+
 
 /**
  * HTTP AWS Authentication handler
@@ -160,7 +161,7 @@ class AWS extends AbstractAuth {
      */
     protected function validateRFC2616Date($dateHeader) {
 
-        $date = Util::parseHTTPDate($dateHeader);
+        $date = HTTP\parseDate($dateHeader);
 
         // Unknown format
         if (!$date) {
