+++ tests/HTTP/NegotiateTest.php
@@ -2,76 +2,7 @@
 
 namespace Sabre\HTTP;
 
-class UtilTest extends \PHPUnit_Framework_TestCase {
-
-    function testParseHTTPDate() {
-
-        $times = [
-            'Wed, 13 Oct 2010 10:26:00 GMT',
-            'Wednesday, 13-Oct-10 10:26:00 GMT',
-            'Wed Oct 13 10:26:00 2010',
-        ];
-
-        $expected = 1286965560;
-
-        foreach ($times as $time) {
-            $result = Util::parseHTTPDate($time);
-            $this->assertEquals($expected, $result->format('U'));
-        }
-
-        $result = Util::parseHTTPDate('Wed Oct  6 10:26:00 2010');
-        $this->assertEquals(1286360760, $result->format('U'));
-
-    }
-
-    function testParseHTTPDateFail() {
-
-        $times = [
-            //random string
-            'NOW',
-            // not-GMT timezone
-            'Wednesday, 13-Oct-10 10:26:00 UTC',
-            // No space before the 6
-            'Wed Oct 6 10:26:00 2010',
-            // Invalid day
-            'Wed Oct  0 10:26:00 2010',
-            'Wed Oct 32 10:26:00 2010',
-            'Wed, 0 Oct 2010 10:26:00 GMT',
-            'Wed, 32 Oct 2010 10:26:00 GMT',
-            'Wednesday, 32-Oct-10 10:26:00 GMT',
-            // Invalid hour
-            'Wed, 13 Oct 2010 24:26:00 GMT',
-            'Wednesday, 13-Oct-10 24:26:00 GMT',
-            'Wed Oct 13 24:26:00 2010',
-        ];
-
-        foreach ($times as $time) {
-            $this->assertFalse(Util::parseHTTPDate($time), 'We used the string: ' . $time);
-        }
-
-    }
-
-    function testTimezones() {
-
-        $default = date_default_timezone_get();
-        date_default_timezone_set('Europe/Amsterdam');
-
-        $this->testParseHTTPDate();
-
-        date_default_timezone_set($default);
-
-    }
-
-    function testToHTTPDate() {
-
-        $dt = new \DateTime('2011-12-10 12:00:00 +0200');
-
-        $this->assertEquals(
-            'Sat, 10 Dec 2011 10:00:00 GMT',
-            Util::toHTTPDate($dt)
-        );
-
-    }
+class NegotiateTest extends \PHPUnit_Framework_TestCase {
 
     /**
      * @dataProvider negotiateData
@@ -80,7 +11,7 @@ class UtilTest extends \PHPUnit_Framework_TestCase {
 
         $this->assertEquals(
             $expected,
-            Util::negotiate($acceptHeader, $available)
+            negotiateContentType($acceptHeader, $available)
         );
 
     }
